utils: reorg
