+++ src/client.ts
@@ -18,7 +18,8 @@ import {
 import { getJSON, setJSON } from "./skydb";
 import { getEntry, getEntryUrl, setEntry } from "./registry";
 
-import { addUrlQuery, defaultPortalUrl, makeUrl } from "./utils";
+import { defaultPortalUrl } from "./utils/skylink";
+import { addUrlQuery, makeUrl } from "./utils/url";
 
 /**
  * Custom client options.
