+++ src/index.ts
@@ -1,16 +1,15 @@
 export { SkynetClient } from "./client";
 export { deriveChildSeed, genKeyPairAndSeed, genKeyPairFromSeed } from "./crypto";
+export { getRelativeFilePath, getRootDirectory } from "./utils/file";
+export { MAX_REVISION } from "./utils/number";
 export {
-  MAX_REVISION,
   defaultPortalUrl,
-  defaultSkynetPortalUrl,
-  getRelativeFilePath,
-  getRootDirectory,
   parseSkylink,
   uriHandshakePrefix,
   uriHandshakeResolverPrefix,
   uriSkynetPrefix,
-} from "./utils";
+} from "./utils/skylink";
+export { defaultSkynetPortalUrl } from "./utils/url";
 
 // Export types.
 
@@ -20,4 +19,4 @@ export type { CustomDownloadOptions, ResolveHnsResponse } from "./download";
 export type { CustomGetEntryOptions, CustomSetEntryOptions, SignedRegistryEntry, RegistryEntry } from "./registry";
 export type { CustomGetJSONOptions, CustomSetJSONOptions, VersionedEntryData } from "./skydb";
 export type { CustomUploadOptions, UploadRequestResponse } from "./upload";
-export type { ParseSkylinkOptions } from "./utils";
+export type { ParseSkylinkOptions } from "./utils/skylink";
