+++ src/utils/url.ts
@@ -1,0 +1,45 @@
+import parse from "url-parse";
+import urljoin from "url-join";
+import { trimSuffix } from './string';
+
+export const defaultSkynetPortalUrl = "https://siasky.net";
+
+/**
+ * Adds a subdomain to the given URL.
+ *
+ * @param url - The URL.
+ * @param subdomain - The subdomain to add.
+ * @returns - The final URL.
+ */
+ export function addSubdomain(url: string, subdomain: string): string {
+  const urlObj = new URL(url);
+  urlObj.hostname = `${subdomain}.${urlObj.hostname}`;
+  const str = urlObj.toString();
+  return trimSuffix(str, "/");
+}
+
+/**
+ * Adds a query to the given URL.
+ *
+ * @param url - The URL.
+ * @param query - The query parameters.
+ * @returns - The final URL.
+ */
+ export function addUrlQuery(url: string, query: Record<string, unknown>): string {
+  const parsed = parse(url, true);
+  // Combine the desired query params with the already existing ones.
+  query = { ...parsed.query, ...query };
+  parsed.set("query", query);
+  return parsed.toString();
+}
+
+/**
+ * Properly joins paths together to create a URL. Takes a variable number of
+ * arguments.
+ *
+ * @param args - Array of URL parts to join.
+ * @returns - Final URL constructed from the input parts.
+ */
+ export function makeUrl(...args: string[]): string {
+  return args.reduce((acc, cur) => urljoin(acc, cur));
+}
